+++ CHANGES.md
@@ -1,9 +1,29 @@
 ## v2.2.0 (2019-08-11)
 
+* Mark trunk as 4.12.0 and add the 4.11 as a development beta (@avsm).
+
+## v2.5.0 (2020-03-22)
+
+* Add OCaml 4.09.1 and 4.10.0 releases (@avsm @kit-ty-kate).
+* Add `equal` function to test versions for equality (#7 @kit-ty-kate)
+* Add a top-level printer hint for `Ocaml_version.t` (@avsm)
+
+## v2.4.0 (2020-02-20)
+
+* Add `unreleased_betas` list to enumerate the latest release-candidate
+  version for each unreleased minor OCaml series (#4 @craigfe).
+* Do not `open Result` to clean up internal code (#5 @avsm).
+
+## v2.3.0 (2019-08-29)
+
+* Support OCaml 4.02.3, which brings back a dependency on
+  the `result` library.
+
+## v2.2.0
+
 * Add OCaml 4.08.1 release
 
 ## v2.1.0 (2019-06-26)
-
 * Add OCaml 4.08.0 release
 * Add support for 4.10 as the new trunk.